requirements.txt
streamlit
pandas
plotly
pillow
kaleido
